+++ pylatex/quantities.py
@@ -1,8 +1,12 @@
 # -*- coding: utf-8 -*-
 """
-This module implements the classes that deals with quantities objects.
+This module implements classes that deal with quantities.
 
-It requires the latex package SIunitx.
+It converts the objects from the quantities package to latex strings that
+display them using the SIunitx package. Not all units work because of name
+differences between quantities and SIunitx. If you find one that doesn't work
+please create a pull request that adds it to the ``UNIT_NAME_TRANSLATIONS``
+dictionary.
 
 ..  :copyright: (c) 2015 by Björn Dahlgren.
     :license: MIT, see License for more details.
@@ -15,7 +19,15 @@ from .package import Package
 from .utils import NoEscape, escape_latex
 
 
+# Translations for names used in the quantities package to ones used by SIunitx
+UNIT_NAME_TRANSLATIONS = {
+    'Celsius': 'celsius',
+}
+
+
 def _dimensionality_to_siunitx(dim):
+    import quantities as pq
+
     string = ''
     items = dim.items()
     for unit, power in sorted(items, key=itemgetter(1), reverse=True):
@@ -26,7 +38,26 @@ def _dimensionality_to_siunitx(dim):
             continue
         else:
             substring = ''
-        substring += '\\' + unit.name
+
+        prefixes = [x for x in dir(pq.prefixes) if not x.startswith('_')]
+        for prefix in prefixes:
+            # Split unitname into prefix and actual name if possible
+            if unit.name.startswith(prefix):
+                substring += '\\' + prefix
+                name = unit.name[len(prefix)]
+                break
+        else:
+            # Otherwise simply use the full name
+            name = unit.name
+
+        try:
+            # Check if the name is different in SIunitx
+            name = UNIT_NAME_TRANSLATIONS[name]
+        except KeyError:
+            pass
+
+        substring += '\\' + name
+
         if power > 1:
             substring += r'\tothe{' + str(power) + '}'
         string += substring
@@ -55,20 +86,20 @@ class Quantity(Command):
         >>> import quantities as pq
         >>> speed = 3.14159265 * pq.meter / pq.second
         >>> Quantity(speed, options={'round-precision': 3,
-        ...                          'round-mode': 'figures'})
+        ...                          'round-mode': 'figures'}).dumps()
         '\\SI[round-mode=figures,round-precision=3]{3.14159265}{\meter\per\second}'
 
         Uncertainties are also handled:
 
         >>> length = pq.UncertainQuantity(16.0, pq.meter, 0.3)
         >>> width = pq.UncertainQuantity(16.0, pq.meter, 0.4)
-        >>> Quantity(length*width)
+        >>> Quantity(length*width).dumps()
         '\\SI{256.0 +- 0.5}{\meter\tothe{2}}
 
         Ordinary numbers are also supported:
 
         >>> Avogadro_constant = 6.022140857e23
-        >>> Quantity(Avogadro_constant, options={'round-precision': 3})
+        >>> Quantity(Avogadro_constant, options={'round-precision': 3}).dumps()
         '\\num[round-precision=3]{6.022e23}'
 
         """